+++ R/utils.R
@@ -69,6 +69,8 @@ column_style_check <- function(columns) {
 }
 
 check_column_type <- function(x) {
+  if (inherits(x, c('numFormat', 'dateFormatter'))) return(invisible())
+
   av_options <-
     c(
       "text",
@@ -97,11 +99,36 @@ update_col_list_with_classes <- function(data, col_list) {
   is_testing <- is_testing()
 
   for (col_name in names(col_classes)) {
+    if (inherits(col_list[[col_name]]$columnType, 'numFormat')) {
+      stopifnot(
+        "'number_format()' can only be applied to a numeric column type" = col_classes[[
+          col_name
+        ]] %in%
+          c("numeric", "integer")
+      )
+      col_list[[col_name]]$columnType <- unclass(
+        col_list[[col_name]]$columnType
+      )
+      col_list[[col_name]]$columnType$initializeNumFormat <- TRUE
+    }
+
+    if (inherits(col_list[[col_name]]$columnType, 'dateFormatter')) {
+      stopifnot(
+        "'date_formatter()' can only be applied to date component" = col_classes[[
+          col_name
+        ]][1] %in%
+          c("Date", "POSIXct", "POSIXlt")
+      )
+      col_list[[col_name]]$columnType <-
+        unclass(col_list[[col_name]]$columnType)
+      col_list[[col_name]]$columnType$initializeDateFormat <- TRUE
+    }
+
     if (is.null(col_list[[col_name]]$columnType)) {
-      if (col_classes[[col_name]] %in% c("numeric", "integer")) {
+      if (col_classes[[col_name]][1] %in% c("numeric", "integer")) {
         col_list[[col_name]]$columnType <- "number"
       } else if (
-        col_classes[[col_name]] == "factor" && any(in_shiny, is_testing)
+        col_classes[[col_name]][1] == "factor" && any(in_shiny, is_testing)
       ) {
         # This is to recover the possible choices for a factor column.
         menu_opt <- lapply(
@@ -181,3 +208,22 @@ make_table_editable <- function(columns, editable = FALSE) {
 }
 
 `%notin%` <- Negate('%in%')
+
+get_bcp47_locale <- function() {
+  locale <- Sys.getlocale("LC_CTYPE")
+
+  # Extract language and region (e.g., from "en_US.UTF-8")
+  match <- regexec("^([a-zA-Z]{2})[_-]([a-zA-Z]{2})", locale)
+  parts <- regmatches(locale, match)[[1]]
+
+  if (length(parts) >= 3) {
+    lang <- tolower(parts[2])
+    region <- toupper(parts[3])
+    return(paste0(lang, "-", region)) # BCP 47 format: "en-US"
+  } else {
+    warning(
+      "Could not parse locale in BCP 47 format; returning fallback 'en-US'"
+    )
+    return("en-US")
+  }
+}